move csv parsing into separate component
